refactor(seed): give the warehouse write path its own module
`generators/base.py` held two nouns: the row-shaping helpers every
generator draws from — calendar, deterministic RNG, pickers — and the
path that writes generated rows into ClickHouse. The second grew past the
first when the insert gained its reconciliation, taking the file to 451
lines, past the size AGENTS.md asks a module to split at.

`generators/insert.py` now owns the write path: RESET_TARGETS, truncate,
the target shape, the plan, and the insert itself. base.py keeps what its
name says and drops to 155 lines. The dependency runs one way — the write
path reads `deterministic_uuid` from base, and nothing reads back.

No behaviour change: the same functions, moved, with their importers
repointed.
