+++ src/ingestion/tools/seed/insight_seed/generators/ai.py
@@ -17,16 +17,15 @@ from typing import TYPE_CHECKING
 from ..profiles import TEAM_PROFILES, Person
 from .base import (
     anchor_datetime,
-    bulk_insert,
     days_window,
     deterministic_int,
     deterministic_uuid,
     persona_multiplier,
     poisson,
     seeded_rng,
-    truncate,
     weekday_multiplier,
 )
+from .insert import bulk_insert, truncate
 
 if TYPE_CHECKING:
     import clickhouse_connect.driver.client
