+++ src/ingestion/tools/seed/insight_seed/generators/task.py
@@ -25,15 +25,14 @@ from ..profiles import TEAM_PROFILES, Person
 from .base import (
     anchor_date,
     anchor_datetime,
-    bulk_insert,
     days_window,
     deterministic_uuid,
     persona_multiplier,
     poisson,
     seeded_rng,
-    truncate,
     weekday_multiplier,
 )
+from .insert import bulk_insert, truncate
 
 if TYPE_CHECKING:
     import clickhouse_connect.driver.client
