+++ src/ingestion/tools/seed/tests/test_bulk_insert.py
@@ -18,7 +18,7 @@ import datetime as _dt
 
 import pytest
 
-from insight_seed.generators import base
+from insight_seed.generators import insert
 
 _ROWS: list[tuple[object, ...]] = [("p1", 3), ("p2", 5)]
 
@@ -29,8 +29,8 @@ def _shape(
     engine: str = "MergeTree",
     sorting_key: str = "",
     key_columns: tuple[str, ...] = (),
-) -> base.TargetShape:
-    return base.TargetShape(
+) -> insert.TargetShape:
+    return insert.TargetShape(
         columns=columns,
         key_columns=key_columns,
         engine=engine,
@@ -40,7 +40,7 @@ def _shape(
 
 def _collapsing(
     columns: dict[str, str], key: str, key_columns: tuple[str, ...]
-) -> base.TargetShape:
+) -> insert.TargetShape:
     return _shape(columns, engine="ReplacingMergeTree", sorting_key=key, key_columns=key_columns)
 
 
@@ -51,14 +51,14 @@ _KEY_SHAPE = _collapsing(
 )
 
 
-def _plan_over_key_shape() -> base.InsertPlan:
-    return base.plan_insert(
+def _plan_over_key_shape() -> insert.InsertPlan:
+    return insert.plan_insert(
         "silver", "class_git_commits", ["person_id", "commits"], _ROWS, _KEY_SHAPE
     )
 
 
 def test_a_column_the_target_lacks_is_dropped_with_its_values() -> None:
-    plan = base.plan_insert(
+    plan = insert.plan_insert(
         "silver",
         "class_git_commits",
         ["person_id", "commits", "tool_label"],
@@ -73,11 +73,11 @@ def test_a_column_the_target_lacks_is_dropped_with_its_values() -> None:
 
 def test_a_target_with_no_writable_column_is_refused() -> None:
     with pytest.raises(RuntimeError, match="class_absent"):
-        base.plan_insert("silver", "class_absent", ["person_id"], _ROWS, _shape({}))
+        insert.plan_insert("silver", "class_absent", ["person_id"], _ROWS, _shape({}))
 
 
 def test_what_the_generator_leaves_to_the_engine_is_named() -> None:
-    plan = base.plan_insert(
+    plan = insert.plan_insert(
         "silver",
         "class_git_commits",
         ["person_id"],
@@ -89,7 +89,7 @@ def test_what_the_generator_leaves_to_the_engine_is_named() -> None:
 
 
 def test_a_day_grain_date_widens_to_aware_utc_midnight() -> None:
-    plan = base.plan_insert(
+    plan = insert.plan_insert(
         "silver",
         "class_git_commits",
         ["person_id", "day"],
@@ -117,8 +117,8 @@ def test_an_omitted_key_column_is_refused_where_rows_collapse(engine: str, colla
         key_columns=("person_id", "day"),
     )
 
-    def plan() -> base.InsertPlan:
-        return base.plan_insert(
+    def plan() -> insert.InsertPlan:
+        return insert.plan_insert(
             "silver", "class_git_commits", ["person_id", "commits"], _ROWS, shape
         )
 
@@ -140,7 +140,7 @@ def test_a_column_the_key_reads_through_an_expression_is_still_demanded() -> Non
     )
 
     with pytest.raises(RuntimeError, match="ts") as caught:
-        base.plan_insert("silver", "class_git_commits", ["person_id"], [("p1",), ("p2",)], shape)
+        insert.plan_insert("silver", "class_git_commits", ["person_id"], [("p1",), ("p2",)], shape)
 
     assert "toDate(ts)" in str(caught.value), "the refusal quotes the key it is about"
 
@@ -165,7 +165,7 @@ def test_the_same_rows_seed_to_the_same_keys() -> None:
 
 
 def test_a_fixed_width_key_column_is_filled_to_its_width() -> None:
-    plan = base.plan_insert(
+    plan = insert.plan_insert(
         "silver",
         "class_git_commits",
         ["person_id"],
@@ -190,7 +190,7 @@ def test_a_version_bump_alone_does_not_change_the_key() -> None:
         ("unique_key",),
     )
     keys = [
-        base.plan_insert(
+        insert.plan_insert(
             "silver", "class_git_commits", ["person_id", "_version"], [("p1", version)], shape
         ).rows[0][2]
         for version in (1, 2)
@@ -233,7 +233,7 @@ class _RecordingClient:
 def test_the_plan_reaches_the_driver_with_its_table_and_database() -> None:
     client = _RecordingClient()
 
-    written = base.bulk_insert(
+    written = insert.bulk_insert(
         client,  # type: ignore[arg-type]
         "silver",
         "class_git_commits",
@@ -248,7 +248,7 @@ def test_the_plan_reaches_the_driver_with_its_table_and_database() -> None:
 def test_an_empty_row_set_reaches_no_server_at_all() -> None:
     client = _RecordingClient()
 
-    assert base.bulk_insert(client, "silver", "class_git_commits", ["person_id"], []) == 0  # type: ignore[arg-type]
+    assert insert.bulk_insert(client, "silver", "class_git_commits", ["person_id"], []) == 0  # type: ignore[arg-type]
     assert (client.queries, client.inserted) == ([], [])
 
 
@@ -258,7 +258,7 @@ def test_the_shape_holds_only_what_a_generator_can_write_and_what_the_key_reads(
     supply it. Which columns count towards the key is the server's answer too."""
     client = _RecordingClient()
 
-    shape = base._target_shape(client, "silver", "class_git_commits")  # type: ignore[arg-type]
+    shape = insert._target_shape(client, "silver", "class_git_commits")  # type: ignore[arg-type]
 
     assert shape.key_columns == ("person_id",)
     assert "default_kind NOT IN ('MATERIALIZED', 'ALIAS')" in client.columns_query()
